Update main.py
Tied up downloading stock/asset from yfinance. Once downloaded, the downloaded dataframe needs to be formatted to gym-trader-env format. It them creates two different dataframes, one for training, another for testing. The training environment is initialized with the training dataframe. A PPO model is created and saved. It then uses the model and runs in testing environment and outputs the result.

*RecurrentPPO doesn't work on Mac, so I use PPO. You can switch back and forth depending on your platform.